+++ BussinessObject/Models/Payment.cs
@@ -11,8 +11,8 @@ namespace BussinessObject.Models
     public class Payment
     {
         [Key]
-        public int paymentId { get; set; }
-        public int paymentAmount { get; set; }
+        public long paymentId { get; set; }
+        public Double paymentAmount { get; set; }
         public DateTime paymentDate { get; set; }
         [ForeignKey("appointmentId")]
         public Appointment appointment_Id { get; set; }
